requirements.txt
dj-database-url==0.4.1
Django==1.9.7
djangorestframework==3.6.3
gunicorn==19.6.0
psycopg2==2.6.2
whitenoise==2.0.6
Pillow==4.1.1
